Split partition processor manager module up into smaller modules
Split PartitionProcessorManager into separate modules

The partition_processor_manager.rs has grown quite a bit. This commit tries
to make it more manageable by splitting some of its components into their own
modules.
